requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
textblob==0.17.1
chromadb==0.4.18
httpx==0.25.2
vaderSentiment==3.3.2
scikit-learn==1.3.2
numpy==1.24.3
psycopg2-binary==2.9.9